requirements.txt
numpy~=1.24.1
torch~=2.1.1+cu118
dgl~=1.1.2+cu118
tqdm~=4.66.1
pyyaml~=6.0.1